Fixed issue with ssl deduplication
Due to ssl deduplication in the event of parameters that can be
vectored with ssls, it can happen that there are 2 ssl with the
same inx. If so there will be double free if these are placed both
in the freeable ssls.  So do not put duplicate inxes in freeable
ssls.