Fixed issue checking geo pred with given geo box.
When checking geo pred with given geo box, read the geo boxes from
the rdf obj table.  When one is missing there are less rows of out
than in.  Index the ser dc, the result from the table with the pos
of the lc. Will not go out of range.  Get the other geo from the
set of input to which the set of output corresponds