Fixed issue with sn_group and sn_person tables
Since V7 does not support that we now create sn_entity as a view
onto sn_person and sn_group. This, however, requires some updating
if we have existing data.  To that end we rename the old tables,
the create the new ones how we want them, copy the old data, and
finally remove the old tables.