fix Virtuoso compilation issue on Solaris while compiling with GNU C
This patch fixes compilation issue on Solaris while using GNU C
for compilation. Although this is not officially supported compiler
for this OS still it is good to have it as a fall back option in case
something's going wrong and you'd like to know if this is Solaris related
issue or compiler related one.
The fix is easy, Solaris does not provide NAN nor INFINITY defines so
we're using those provided for WIN32. Please note that this is case
only when compiling with GNU C. Sun's own C compiler supports both defines
of course!