Fix the non-fatal issue about the operand of the `sizeof`.
Result of `s_realloc`, i.e. `realloc` is converted to `char**`, which is incompatiable with the sizeof operand type `char**`.

Given the code below, the operand of sizeof is `int`, the allocated memory type will be `int*`. In the same way, if the operand of sizeof is `char**`, the allocated memory type will be `char***`.
```
int *ptr = (int*)malloc(sizeof(int));
```

The operand of sizeof, whether `char*` or `char**`, result in the same program behavior. But use the consistent manner will improve the readability.