Added support for more flexible buffer configuration
Traditionally we provide a number of precalculated configurations
in the virtuoso.ini, based on the amount of free memory in the
system, e.g.:

  ;; Uncomment next two lines if there is 8 GB system memory free
  NumberOfBuffer  = 680000
  MaxDirtyBuffers = 500000

The user has to manually remove comments from the appropriate lines
and not forget to modify the values when moving to a bigger/smaller
system.

If the user forgets to modify the virtuoso.ini, the defaults will
take very little memory, but will result in very poor performance.

Most users do not know how to calculate the amount of memory used
by the NumberOfBufffers which uses the following formula:

  amount memory allocated = NumberOfBuffers * 8192 bytes

This makes it difficult for users to tune Virtuoso memory consumption.

To solve this problem, NumberOfBuffers can now be expressed as an
amount of memory, e.g.:

  NumberofBuffers = 5000M      ; calculate max NumberOfBuffers that will fit in 5000MB mem
  NumberOfBuffers = 5G         ; calculate max NumberOfBuffers that will fit in 5GB mem

When using EC2, Azure, or containers like kubernetes and docker,
we need a more flexible configuration, which allows for a single
Virtuoso configuration to scale to the amount of memory installed
on the machine.

To scale automatically, NumberOfBuffers can now be expressed as a
percentage, e.g.:

  NumberOfBuffers = 50%        ; use 50% of installed memory for buffers

This will allow OpenLink to create new marketplace instances for
BYOL, PAYGO and DBpedia instances that will automatically scale
when the enduser instantiates it on a 16GB, or a 64GB or larger
system.

Additionally we now allow MaxDirtyBuffers to be expressed as a
percentage of the NumberOfBuffers that can be dirty.

  MaxDirtyBuffers = 75%        ; allow upto 75% of NumberOfBuffers to be dirty