Fixed issue with very long sparql queries
If we have a very long query in the textarea, the url will be larger
than 14000 chars. Reportedly most modern browsers seem to have a problem
with this when submitting using a GET method on the form.

Although we have a registry setting called 'sparql-ui-max-url' that can be
set as a maximum length for switching to a POST method on the form,
this is off by default.

This patch makes sure that such large queries will be automatically be
submitted using the POST method.